Refactor distance angle regulator
