+++ mep3_navigation/src/distance_angle/distance_angle_regulator.cpp
@@ -1,4 +1,4 @@
-#include "distance_angle/distance_angle_regulator.hpp"
+#include "mep3_navigation/distance_angle/distance_angle_regulator.hpp"
 #include <iostream>
 
 #include "tf2/LinearMath/Quaternion.h"
@@ -6,15 +6,13 @@
 #include <cmath>
 #include <string>
 
-
-
 DistanceAngleRegulator::DistanceAngleRegulator() : Node("distance_angle_regulator")
 {
     odometry_subscription_ = this->create_subscription<nav_msgs::msg::Odometry>(
         "/odom", 100, std::bind(&DistanceAngleRegulator::odometry_callback, this, _1));
     twist_publisher_ = this->create_publisher<geometry_msgs::msg::Twist>("/cmd_vel", 10);
 
-    command_subscription_ = this->create_subscription<distance_angle_interfaces::msg::MotionCommand>("/da_command", 100, std::bind(&DistanceAngleRegulator::command_callback, this, _1));
+    command_subscription_ = this->create_subscription<mep3_msgs::msg::MotionCommand>("/da_command", 100, std::bind(&DistanceAngleRegulator::command_callback, this, _1));
 
     this->declare_parameter("kp_distance", 25000.0);
     this->declare_parameter("ki_distance", 1.5);
@@ -45,7 +43,7 @@ DistanceAngleRegulator::DistanceAngleRegulator() : Node("distance_angle_regulato
 
     robot_distance_ = 0;
     position_initialized_ = false;
-    
+
     distance_profile_ = MotionProfile(0, 0.1, 0.02);
     angle_profile_ = MotionProfile(0, 0.5, 0.25);
 }
@@ -58,7 +56,7 @@ void DistanceAngleRegulator::odometry_callback(const nav_msgs::msg::Odometry::Sh
     double robot_y = msg->pose.pose.position.y;
 
     if (!position_initialized_)
-    {   
+    {
         prev_robot_x_ = robot_x;
         prev_robot_y_ = robot_y;
         position_initialized_ = true;
@@ -72,16 +70,16 @@ void DistanceAngleRegulator::odometry_callback(const nav_msgs::msg::Odometry::Sh
         double roll, pitch, yaw;
         m.getRPY(roll, pitch, yaw);
         robot_angle_ = yaw;
-        
+
         distance_profile_.plan(0, 0, 0, 0, time);
         angle_profile_.plan(robot_angle_, robot_angle_, 0, 0, time);
     }
 
     tf2::Quaternion q(
-            msg->pose.pose.orientation.x,
-            msg->pose.pose.orientation.y,
-            msg->pose.pose.orientation.z,
-            msg->pose.pose.orientation.w);
+        msg->pose.pose.orientation.x,
+        msg->pose.pose.orientation.y,
+        msg->pose.pose.orientation.z,
+        msg->pose.pose.orientation.w);
     tf2::Matrix3x3 m(q);
     double roll, pitch, yaw;
     m.getRPY(roll, pitch, yaw);
@@ -94,9 +92,8 @@ void DistanceAngleRegulator::odometry_callback(const nav_msgs::msg::Odometry::Sh
     int sign = 1;
     if (std::abs(robot_angle_ - distance_increment_angle) > 0.1)
         sign = -1;
-    
-    robot_distance_ += sign * distance_increment;
 
+    robot_distance_ += sign * distance_increment;
 
     regulator_distance_.feedback = robot_distance_;
     regulator_angle_.feedback = robot_angle_;
@@ -107,7 +104,6 @@ void DistanceAngleRegulator::odometry_callback(const nav_msgs::msg::Odometry::Sh
     double normalized_angle_error = angle_normalize(angle_ref - robot_angle_);
     regulator_angle_.reference = regulator_angle_.feedback + normalized_angle_error;
 
-    
     pid_regulator_update(&regulator_distance_);
     pid_regulator_update(&regulator_angle_);
 
@@ -124,7 +120,7 @@ void DistanceAngleRegulator::odometry_callback(const nav_msgs::msg::Odometry::Sh
         std::cout << "Regulator angle error deg: " << regulator_angle_.error * 180 / M_PI << std::endl;
         std::cout << std::endl;
     }
-    
+
     geometry_msgs::msg::Twist motor_command;
     motor_command.linear.x = regulator_distance_.command / 10000.0;
     motor_command.angular.z = regulator_angle_.command / 10000.0;
@@ -132,7 +128,6 @@ void DistanceAngleRegulator::odometry_callback(const nav_msgs::msg::Odometry::Sh
 
     prev_robot_x_ = robot_x;
     prev_robot_y_ = robot_y;
-
 }
 
 rcl_interfaces::msg::SetParametersResult DistanceAngleRegulator::parameters_callback(const std::vector<rclcpp::Parameter> &parameters)
@@ -141,7 +136,7 @@ rcl_interfaces::msg::SetParametersResult DistanceAngleRegulator::parameters_call
     result.successful = true;
     result.reason = "success";
 
-    for (const auto &param: parameters)
+    for (const auto &param : parameters)
     {
         if (param.get_name() == "kp_distance" && param.get_type() == rclcpp::ParameterType::PARAMETER_DOUBLE)
         {
@@ -176,12 +171,12 @@ rcl_interfaces::msg::SetParametersResult DistanceAngleRegulator::parameters_call
     return result;
 }
 
-void DistanceAngleRegulator::command_callback(const distance_angle_interfaces::msg::MotionCommand::SharedPtr msg)
+void DistanceAngleRegulator::command_callback(const mep3_msgs::msg::MotionCommand::SharedPtr msg)
 {
     if (msg->command == "forward")
     {
         forward(msg->value);
-    }    
+    }
     else if (msg->command == "rotate_absolute")
     {
         rotate_absolute(msg->value);
@@ -207,24 +202,24 @@ double DistanceAngleRegulator::angle_normalize(double angle)
 }
 void DistanceAngleRegulator::forward(double distance)
 {
-    rclcpp::Time time = this->get_clock()->now();   
+    rclcpp::Time time = this->get_clock()->now();
     distance_profile_.plan(distance_profile_.get_position(), distance_profile_.get_position() + distance, distance_profile_.get_velocity(), 0, time);
 }
 
 void DistanceAngleRegulator::rotate_absolute(double angle)
 {
     angle = angle_normalize(angle);
-    rclcpp::Time time = this->get_clock()->now();   
+    rclcpp::Time time = this->get_clock()->now();
     angle_profile_.plan(robot_angle_, angle, angle_profile_.get_velocity(), 0, time);
 }
 
 void DistanceAngleRegulator::rotate_relative(double angle)
 {
-    rclcpp::Time time = this->get_clock()->now();   
+    rclcpp::Time time = this->get_clock()->now();
     angle_profile_.plan(angle_profile_.get_position(), angle_profile_.get_position() + angle, angle_profile_.get_velocity(), 0, time);
 }
 
-int main(int argc, char * argv[])
+int main(int argc, char *argv[])
 {
     rclcpp::init(argc, argv);
     rclcpp::spin(std::make_shared<DistanceAngleRegulator>());
