--- mep3_tools/README.md
@@ -1,2 +1,0 @@
-# Tools
-Various helper scripts and libraries
