--- mep3_tools/trapezoidal_trajectory_generator/.idea/.gitignore
@@ -1,3 +1,0 @@
-# Default ignored files
-/shelf/
-/workspace.xml
